app/requirements.txt
Flask==3.0.0
boto3==1.34.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-exporter-otlp==1.21.0
